Separate module for visual voice controls
